fix: upgrade button wiring
