+++ src/components/frame/Extensions/pages/partials/ActionButtons.tsx
@@ -1,14 +1,6 @@
 import React, { FC } from 'react';
 import { useIntl } from 'react-intl';
-import {
-  ColonyRole,
-  ColonyVersion,
-  Extension,
-  ExtensionVersion,
-  Id,
-  isExtensionCompatible,
-} from '@colony/colony-js';
-import { toast } from 'react-toastify';
+import { ColonyRole, Id } from '@colony/colony-js';
 import { useLocation, useNavigate } from 'react-router-dom';
 import { useFormContext } from 'react-hook-form';
 
@@ -27,14 +19,10 @@ import { ActionButtonProps } from './types';
 import HeadingIcon from './HeadingIcon';
 import ExtensionStatusBadge from '~v5/common/Pills/ExtensionStatusBadge';
 import ActiveInstalls from './ActiveInstalls';
-import { mapPayload } from '~utils/actions';
-import { ActionButton } from '~shared/Button';
-import { ActionTypes } from '~redux';
-import Toast from '~shared/Extensions/Toast/Toast';
-import { ButtonProps } from '~v5/shared/Button/types';
 import InstallButton from './InstallButton';
 import { COLONY_EXTENSION_SETUP_ROUTE } from '~routes';
 import { addressHasRoles } from '~utils/checks';
+import UpgradeButton from './UpgradeButton';
 
 const displayName = 'frame.Extensions.pages.partials.ActionButtons';
 
@@ -46,24 +34,7 @@ const ActionButtons: FC<ActionButtonProps> = ({
   extensionStatusText,
 }) => {
   const { user } = useAppContext();
-  const { colony, isSupportedColonyVersion } = useColonyContext();
-
-  const { colonyAddress = '' } = colony || {};
-
-  const transformUpgrade = mapPayload(() => ({
-    colonyAddress: colony?.colonyAddress,
-    extensionId: extensionData.extensionId,
-    version: extensionData.availableVersion,
-  }));
-
-  const extensionCompatible = isExtensionCompatible(
-    Extension[extensionData.extensionId],
-    extensionData.availableVersion as ExtensionVersion,
-    colony?.version as ColonyVersion,
-  );
-
-  const isUpgradeButtonDisabled =
-    !isSupportedColonyVersion || !extensionCompatible;
+  const { colony } = useColonyContext();
 
   const {
     formState: { isValid, isSubmitting },
@@ -159,38 +130,7 @@ const ActionButtons: FC<ActionButtonProps> = ({
       )}
       {isEnableButtonVisible && enableButton}
       {isUpgradeButtonVisible && (
-        <ActionButton<ButtonProps>
-          actionType={ActionTypes.EXTENSION_UPGRADE}
-          values={{ colonyAddress, extensionData }}
-          transform={transformUpgrade}
-          onSuccess={() => {
-            toast.success(
-              <Toast
-                type="success"
-                title={{ id: 'extensionUpgrade.toast.title.success' }}
-                description={{
-                  id: 'extensionUpgrade.toast.description.success',
-                }}
-              />,
-            );
-          }}
-          onError={() => {
-            toast.error(
-              <Toast
-                type="error"
-                title={{ id: 'extensionUpgrade.toast.title.error' }}
-                description={{ id: 'extensionUpgrade.toast.description.error' }}
-              />,
-            );
-          }}
-          button={Button}
-          buttonProps={{
-            mode: 'primarySolid',
-            isFullSize: isMobile,
-            disabled: isUpgradeButtonDisabled,
-            children: formatMessage({ id: 'button.updateVersion' }),
-          }}
-        />
+        <UpgradeButton extensionData={extensionData} />
       )}
     </>
   );
